requirements.txt
django==3.1
djangorestframework==3.11.1
pillow==7.1
pytils
python-decouple
stripe
django-filter
djangorestframework-simplejwt
drf-yasg
psycopg2-binary==2.8.5
gunicorn==20.0.4
whitenoise==5.1.0
dj-database-url==0.5.0
